movendo componentes de hooks para useFiles
